Decomposition module WIP
