+++ doc/modules/classes.rst
@@ -205,17 +205,17 @@ Signal Decomposition
    :toctree: generated/
    :template: class.rst
 
-   pca.PCA
-   pca.ProbabilisticPCA
-   pca.RandomizedPCA
-   fastica.FastICA
-   nmf.NMF
+   decomposition.PCA
+   decomposition.ProbabilisticPCA
+   decomposition.RandomizedPCA
+   decomposition.FastICA
+   decomposition.NMF
 
 .. autosummary::
    :toctree: generated/
    :template: function.rst
 
-   fastica.fastica
+   decomposition.fastica.fastica
 
 Cross Validation
 ================
