+++ examples/plot_ica_vs_pca.py
@@ -33,8 +33,7 @@ print __doc__
 import numpy as np
 import pylab as pl
 
-from scikits.learn.pca import PCA
-from scikits.learn.fastica import FastICA
+from scikits.learn.decomposition import PCA, FastICA
 
 ###############################################################################
 # Generate sample data
