+++ scikits/learn/__init__.py
@@ -36,10 +36,9 @@ except:
 
 
 __all__ = ['cross_val', 'ball_tree', 'cluster', 'covariance', 'datasets',
-           'fastica', 'feature_extraction', 'feature_selection',
+           'decomposition', 'feature_extraction', 'feature_selection',
            'gaussian_process', 'grid_search', 'hmm', 'lda', 'linear_model',
-           'metrics', 'mixture', 'naive_bayes', 'neighbors', 'nmf',
-           'pca', 'pipeline', 'preprocessing', 'qda', 'svm', 'test',
-           'clone', 'pls']
+           'metrics', 'mixture', 'naive_bayes', 'neighbors', 'pipeline', 
+           'preprocessing', 'qda', 'svm', 'test', 'clone', 'pls']
 
 __version__ = '0.8.git'
