+++ scikits/learn/decomposition/__init__.py
@@ -3,4 +3,5 @@ Matrix decomposition algorithms
 """
 
 from .nmf import NMF
-from .pca import PCA, RandomizedPCA, KernelPCA
+from .pca import PCA, RandomizedPCA, ProbabilisticPCA
+from .fastica import FastICA
